+++ sys/src/9/omap/dat.h
@@ -103,7 +103,7 @@ enum
 struct Confmem
 {
 	uintptr	base;
-	usize	npage;
+	ulong	npage;
 	uintptr	limit;
 	uintptr	kbase;
 	uintptr	klimit;
@@ -115,7 +115,7 @@ struct Conf
 	ulong	nproc;		/* processes */
 	Confmem	mem[1];		/* physical memory */
 	ulong	npage;		/* total physical pages of memory */
-	usize	upages;		/* user page pool */
+	ulong	upages;		/* user page pool */
 	ulong	copymode;	/* 0 is copy on write, 1 is copy on reference */
 	ulong	ialloc;		/* max interrupt time allocation in bytes */
 	ulong	pipeqsize;	/* size in bytes of pipe queues */
